moseca/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
antlr4-python3-runtime==4.9.3
    # via omegaconf
anyio==3.7.1
    # via starlette
appdirs==1.4.4
    # via pooch
audioread==3.0.1
    # via librosa
brotli==1.0.9
    # via yt-dlp
certifi==2023.7.22
    # via
    #   requests
    #   yt-dlp
cffi==1.15.1
    # via soundfile
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via uvicorn
cloudpickle==3.1.0
    # via submitit
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   tqdm
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
cython==0.29.36
    # via diffq
decorator==5.1.1
    # via librosa
demucs==4.0.0
    # via -r requirements.in
diffq==0.2.4
    # via
    #   -r requirements.in
    #   demucs
dora-search==0.1.12
    # via
    #   -r requirements.in
    #   demucs
einops==0.6.1
    # via
    #   -r requirements.in
    #   demucs
fastapi==0.100.0
    # via -r requirements.in
filelock==3.16.1
    # via torch
fonttools==4.41.0
    # via matplotlib
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
jinja2==3.1.2
    # via torch
joblib==1.3.1
    # via
    #   librosa
    #   scikit-learn
julius==0.2.7
    # via
    #   -r requirements.in
    #   demucs
kiwisolver==1.4.4
    # via matplotlib
lameenc==1.5.1
    # via
    #   -r requirements.in
    #   demucs
lazy-loader==0.4
    # via librosa
librosa==0.10.0.post2
    # via -r requirements.in
llvmlite==0.40.1
    # via numba
loguru==0.7.0
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.1
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
msgpack==1.0.5
    # via librosa
mutagen==1.46.0
    # via yt-dlp
networkx==3.1
    # via torch
numba==0.57.1
    # via
    #   librosa
    #   resampy
numpy==1.24.4
    # via
    #   -r requirements.in
    #   contourpy
    #   diffq
    #   librosa
    #   matplotlib
    #   numba
    #   openunmix
    #   pandas
    #   resampy
    #   scikit-learn
    #   scipy
    #   soxr
omegaconf==2.3.0
    # via dora-search
openunmix==1.2.1
    # via
    #   -r requirements.in
    #   demucs
packaging==23.1
    # via
    #   lazy-loader
    #   matplotlib
    #   pooch
pandas==1.5.3
    # via -r requirements.in
pillow==10.0.0
    # via
    #   -r requirements.in
    #   matplotlib
pooch==1.6.0
    # via librosa
pycparser==2.22
    # via cffi
pycryptodomex==3.21.0
    # via yt-dlp
pydantic==1.10.9
    # via fastapi
pydub==0.25.1
    # via -r requirements.in
pyparsing==3.1.0
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-multipart==0.0.13
    # via -r requirements.in
pytube==12.1.3
    # via -r requirements.in
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via
    #   demucs
    #   omegaconf
requests==2.31.0
    # via
    #   -r requirements.in
    #   pooch
resampy==0.4.2
    # via -r requirements.in
retrying==1.3.4
    # via dora-search
scikit-learn==1.3.0
    # via librosa
scipy==1.11.1
    # via
    #   -r requirements.in
    #   librosa
    #   scikit-learn
six==1.16.0
    # via
    #   python-dateutil
    #   retrying
sniffio==1.3.0
    # via anyio
soundfile==0.12.1
    # via librosa
soxr==0.3.5
    # via librosa
starlette==0.27.0
    # via fastapi
submitit==1.5.2
    # via dora-search
sympy==1.13.3
    # via torch
threadpoolctl==3.2.0
    # via scikit-learn
torch==2.0.1
    # via
    #   -r requirements.in
    #   demucs
    #   diffq
    #   dora-search
    #   julius
    #   openunmix
    #   torchaudio
torchaudio==2.0.2
    # via
    #   -r requirements.in
    #   demucs
    #   openunmix
tqdm==4.65.0
    # via
    #   demucs
    #   openunmix
treetable==0.2.5
    # via dora-search
typing-extensions==4.7.1
    # via
    #   fastapi
    #   librosa
    #   pydantic
    #   submitit
    #   torch
urllib3==2.0.4
    # via requests
uvicorn==0.22.0
    # via -r requirements.in
websockets==11.0.3
    # via yt-dlp
win32-setctime==1.1.0
    # via loguru
yt-dlp==2023.7.6
    # via -r requirements.in
